requirements.txt
discord
python-dotenv
dalle2
matplotlib
numpy
scipy
web3